+++ GetYourWay/app/src/main/java/test/getyourway/Position.java
@@ -6,16 +6,16 @@ package test.getyourway;
 
 public class Position {
 
-    double x;
-    double y;
+    float x;
+    float y;
 
-    public Position (double x, double y){
+    public Position (float x, float y){
 
         this.x = x;
         this.y = y;
     }
 
-    public void setPosition (double x, double y){
+    public void setPosition (float x, float y){
 
         this.x = x;
         this.y = y;
@@ -26,6 +26,7 @@ public class Position {
         return p;
     }
 }
+
